refactor: move secure security takes in web folder
Ref: #18
